common: move configuration in common
