+++ crates/cli/Cargo.toml
@@ -6,7 +6,6 @@ edition = "2024"
 [dependencies]
 ansi_term = "0.12.1"
 linefeed = "0.6"
-confy = "0.6.1"
 lazy_static = "1.5.0"
 atty = "0.2"
 password-hash = { version = "0.5.0", features = ["getrandom"] }
