+++ crates/cli/src/handlers/user/logout.rs
@@ -1,9 +1,11 @@
-use common::structs::struct_user::{LogoutAnswer, LogoutPost};
+use common::{
+    configuration::loader::{load, write_config},
+    structs::struct_user::{LogoutAnswer, LogoutPost},
+};
 
 use crate::{
     API_URL, TOKEN,
     call::{Answer, call},
-    configuration::loader::{load, write_config},
 };
 
 impl Answer for LogoutAnswer {
