+++ crates/common/Cargo.toml
@@ -6,3 +6,5 @@ edition = "2024"
 [dependencies]
 serde = { version = "1.0.192", features = ["derive"] }
 chrono = { version="0.4.40", features = ["serde"] }
+confy = "0.6.1"
+ansi_term = "0.12.1"
