+++ crates/common/src/lib.rs
@@ -1,1 +1,3 @@
+pub static VERSION: &str = "v3.0.0-dev";
+pub mod configuration;
 pub mod structs;
